pyproject.toml
[tool.poetry]
name = "serial-test"
version = "0.1.0"
description = ""
authors = ["Anton Asieiev <ntn.kiev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
